+++ src/androidTest/java/de/blau/android/resources/TileLayerServerTest.java
@@ -78,7 +78,9 @@ public class TileLayerServerTest {
     @After
     public void teardown() {
         instrumentation.removeMonitor(monitor);
-        context.deleteDatabase(TileLayerDatabase.DATABASE_NAME);
+        if (context != null) {
+            context.deleteDatabase(TileLayerDatabase.DATABASE_NAME);
+        }
         try {
             main.finish();
         } catch (Exception e) {
